Basic entity tracking

+++ bot/phy/inputs.go
@@ -2,7 +2,7 @@ package phy
 
 // Inputs describes the desired movements of the player.
 type Inputs struct {
-	Yaw, Pitch float32
+	Yaw, Pitch float64
 
 	ThrottleX, ThrottleZ float64
 }
